pyproject.toml
[project]
name = "coral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "playwright>=1.49",
    "python-dotenv>=1.0",
]
